pyproject.toml
[tool.poetry]
name = "heltour"
version = "1.0.1"
description = "A [hel]per for [tour]naments.  Intended to run swiss based tournaments, but is actually the source for the lichess4545.com leagues."
authors = ["Lakin Wecker <lakin@structuredabstraction.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
django = "==4.2.*"
django-reversion = "==5.0.*"
psycopg2-binary = "==2.9.9"
#requests[security] = "==2.22.0"
responses = "==0.24.1"
django-bootstrap3 = "==23.*"
gspread = "==0.4.1"
oauth2client = "==2.2.0"
django-ckeditor = "==6.7.0"
django-cacheops = "==7.0.*"
django-contrib-comments = "==2.2.0"
# django until here
django-redis = "==5.4.0"
django-recaptcha = "==4.0.0"
django-static-precompiler = "==2.4"
django-impersonate = "==1.9.1"
django-ipware = "==6.0.0"
redis = "==5.0.1"
websocket-client = "==1.6.4"
celery = "==5.3.6"
icalendar = "==5.0.11"
pillow = "==10.0.1"
pyfcm = "==1.3.1"
letsencrypt = "*"
requests = {extras = ["security"], version = "^2.31.0"}
django-stubs = "4.2.6"
django-stubs-ext = "4.2.5"
django-debug-toolbar = "==4.2.0"
django-select2-forms = "^3.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
# FIXME: these belong here but there appears to be a hard production dependency on them. 
# django-stubs = "1.7.0"
# django-stubs-ext = "^0.3.1"
# django-debug-toolbar = "==3.2.4"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
gunicorn = "==22.0.0"

[tool.poetry.group.sysadmin]
optional = true

[tool.poetry.group.sysadmin.dependencies]
patchwork = "1.0.1"
fabric3 = "==1.14.post1"
baste = { git = "https://bitbucket.org/strabs/baste.git", branch="release-2.2" }

[tool.ruff]
line-length= 88
indent-width = 4
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "strabulous.py",
    "fabfile.py",
    "bin/createteams.py",
    "**/migrations/*.py",
    "sysadmin/*.py",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "heltour.settings"
